backend/healthcare_backend/healthcare_backend/urls.py
"""
Healthcare Project URL Configuration
Main URL routing for the entire project
"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenBlacklistView,
)

urlpatterns = [
    # Django Admin Panel
    path('admin/', admin.site.urls),
    
    # JWT Authentication Endpoints
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/blacklist/', TokenBlacklistView.as_view(), name='token_blacklist'),
    
    # Healthcare API Endpoints (includes all patient/doctor/admin endpoints)
    path('api/', include('healthcare.urls')),
]

# Serve static and media files in development
if settings.DEBUG:
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)

# Customize Django Admin Panel
admin.site.site_header = "Healthcare Management System"
admin.site.site_title = "Healthcare Admin Portal"
admin.site.index_title = "Welcome to Healthcare Administration"
